crud dos livros e estilizando tela de perfil

+++ admin/processDeletePublisher.php
@@ -1,18 +1,26 @@
 <?php
 require_once('../DAO/DAOPublisher.php');
 require_once('../models/Publisher.php');
+require_once('../DAO/DAOPublish.php');
+require_once('../models/Publish.php');
 
 $id = filter_input(INPUT_POST, 'id');
 $return = [];
 
 if($id) {
-    $publisher = new Publisher($id, null, null, null, null);
     $daoPublisher = new DaoPublisher();
-    if($daoPublisher->deletePublisher($publisher)) {
-        $return = ['status' => 'ok'];
-    } else {
-        $return = ['status' => 'sla'];
-    }
+    $daoPublish = new DaoPublish();
+
+    $publisher = new Publisher($id, null, null, null, null);
+    $publish = new Publish($id, null);
+
+    if($daoPublish->deletePubByPubId($publish)) {
+        if($daoPublisher->deletePublisher($publisher)) {
+            $return = ['status' => 'ok'];
+        } else {
+            $return = ['status' => 'sla'];
+        }
+    }   
 } else {
     $return = ['status' => 'erro'];
 }
